fix OTF variables: libOTF uses uint64_t whereas gtg used to give int. Moreover, the upper layer of gtg uses double to store the value, which may cause an overflow when converted into int.
git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/gtg/trunk@128 0c38aafd-2db3-40e8-87a3-ec7d7f65c264

+++ inc/OTF_Structs.h
@@ -98,7 +98,7 @@ typedef struct VariableType {
 typedef struct Variable {
     int parent;
     int type;
-    int value;
+    uint64_t value;
 } Variable_t;
 
 struct otf_color {
